Add fastify as peer dependency
This PR adds fastify as a peer dependency so that it works properly with Yarn v2 without having to override the package. Without it you'll get the following message in your log for every plugin since it is trying to require `fastify/package.json`:
```
fastify not found, proceeding anyway
```
Yarn v2 enforces packages to list all dependencies that are used by that package. In this case, fastify is provided by the user so it should be a peer dependency.